Add Dockerfile for jekyll build
This commit adds a Dockerfile for running the jekyll build and
development server, which should help streamline developer onboarding.
Docker (or Podman, Nerdctl, Finch, ...) is often one of the few tools
found reliably within many systems "trusted computing bases", whereas
`ruby` and `bundle` are not-so-easily available - Fedora's atomic
distributions are such examples of this.

This Dockerfile is, for now, a simple one-stage build atop Debian 12
Bookworm (`bookworm-slim`) as the base image. We lock the parent digest
and use Debian snapshots so as to improve reproducibility of the image
builds.

The Dockerfile performs fairly efficient build layering; for example,
the Debian packages install is independent of sources (other than the
`Dockerfile` itself, of course), and the Ruby packages install will only
run when the `Gemfile` changes.

There are some at-first-glance-odd approaches in the Dockerfile, made to
support running the container conventionally, or when over-mounting the
work-dir to integrate local changes with Jekyll's auto-reload, as per
the `docker` commands added to the README.