Add Dockerfile for jekyll build (#22)
This commit adds a Dockerfile for running the jekyll build and
development server, which should help streamline developer onboarding.
Docker (or Podman, Nerdctl, Finch, ...) is often one of the few tools
found reliably within many systems "trusted computing bases", whereas
`ruby` and `bundle` are not-so-often available - particularly at
compatible versions. Fedora's atomic distributions (Silverblue et al)
are such examples.

This Dockerfile is, for now, a one-stage build atop Debian 12 Bookworm
(`bookworm-slim`) as the base image. We lock the parent digest and use
Debian snapshots to improve reproducibility of the image builds.

The Dockerfile performs fairly efficient build layering; for example,
the Debian packages install is independent of sources (other than the
`Dockerfile` itself, of course), and the Ruby gems install will only run
when the `Gemfile` changes.

The Dockerfile is more intricate than usual, so as to support bi-modal
`jekyll serve` functionality - either when running off static container
sources for conventional `docker run` invocations, or for when
over-mounting the workdir with local sources to leverage Jekyll's
auto-reload functionality.

This commit updates the README with `docker` commands for building and
running the container.